Initial refactoring of HopsanGenerator
- Code cleanup
- Remove GUI parts, they will be restored in a different library later
  It is very strange to launch GUI widgets from an other seemingly non-gui library loaded at runtime.
  Especially when it is loaded into a non-gui application.
