+++ HopsanGenerator/src/GeneratorTypes.cpp
@@ -1,0 +1,314 @@
+/*-----------------------------------------------------------------------------
+
+ Copyright 2017 Hopsan Group
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+
+ The full license is available in the file LICENSE.
+ For details about the 'Hopsan Group' or information about Authors and
+ Contributors see the HOPSANGROUP and AUTHORS files that are located in
+ the Hopsan source code root directory.
+
+-----------------------------------------------------------------------------*/
+
+#include "GeneratorTypes.h"
+
+#include "HopsanEssentials.h"
+#include "Port.h"
+
+
+PortSpecification::PortSpecification(QString porttype, QString nodetype, QString name, bool notrequired, QString defaultvalue)
+{
+    this->porttype = porttype;
+    this->nodetype = nodetype;
+    this->name = name;
+    this->notrequired = notrequired;
+    this->defaultvalue = defaultvalue;
+}
+
+
+ParameterSpecification::ParameterSpecification(QString name, QString displayName, QString description, QString unit, QString init)
+{
+    this->name = name;
+    this->displayName = displayName;
+    this->description = description;
+    this->unit = unit;
+    this->init = init;
+}
+
+
+VariableSpecification::VariableSpecification(QString name, QString init)
+{
+    this->name = name;
+    this->init = init;
+}
+
+
+ComponentSpecification::ComponentSpecification(QString typeName, QString displayName, QString cqsType)
+{
+    this->typeName = typeName;
+    this->displayName = displayName;
+    if(cqsType == "S")
+        cqsType = "Signal";
+    this->cqsType = cqsType;
+
+    this->auxiliaryFunctions = QStringList();
+}
+
+GeneratorNodeInfo::GeneratorNodeInfo(QString nodeType)
+{
+    //! @todo this will only be able to create the default included nodes (which may be a problem in the future)
+    hopsan::HopsanEssentials hopsanCore;
+    hopsan::Node *pNode = hopsanCore.createNode(nodeType.toStdString().c_str());
+    isValidNode = false;
+    if (pNode)
+    {
+        isValidNode = true;
+        niceName = pNode->getNiceName().c_str();
+        for(size_t i=0; i<pNode->getDataDescriptions()->size(); ++i)
+        {
+            const hopsan::NodeDataDescription *pVarDesc = pNode->getDataDescription(i);
+            hopsan::NodeDataVariableTypeEnumT varType = pVarDesc->varType;
+            // Check if  "Q-type variable"
+            if(varType == hopsan::DefaultType || varType == hopsan::FlowType || varType == hopsan::IntensityType)
+            {
+                qVariables << pVarDesc->shortname.c_str();
+                qVariableIds << pVarDesc->id;
+                variableLabels << pVarDesc->name.c_str();
+                varIdx << pVarDesc->id;
+            }
+            // Check if "C-type variable"
+            else if(varType == hopsan::TLMType)
+            {
+                cVariables << pVarDesc->shortname.c_str();
+                cVariableIds << pVarDesc->id;
+                variableLabels << pVarDesc->name.c_str();
+                varIdx << pVarDesc->id;
+            }
+
+        }
+        hopsanCore.removeNode(pNode);
+    }
+}
+
+void GeneratorNodeInfo::getNodeTypes(QStringList &nodeTypes)
+{
+    //! @todo this will only be able to list the default included nodes (which may be a problem in the future)
+    hopsan::HopsanEssentials hopsanCore;
+    std::vector<hopsan::HString> types = hopsanCore.getRegisteredNodeTypes();
+    Q_FOREACH(const hopsan::HString &type, types)
+    {
+        nodeTypes << type.c_str();
+    }
+}
+
+
+InterfacePortSpec::InterfacePortSpec(InterfaceTypesEnumT type, QString component, QString port, QStringList path)
+{
+    this->type = type;
+    this->path = path;
+    this->component = component;
+    this->port = port;
+
+    QStringList inputDataNames;
+    QStringList outputDataNames;
+    QList<size_t> inputDataIds, outputDataIds;
+
+    switch(type)
+    {
+    case InterfacePortSpec::Input:
+        inputDataNames << "";
+        inputDataIds << 0;
+        break;
+    case InterfacePortSpec::Output:
+        outputDataNames << "";
+        outputDataIds << 0;
+        break;
+    case InterfacePortSpec::MechanicQ:
+    {
+        GeneratorNodeInfo gni("NodeMechanic");
+        inputDataNames  << gni.qVariables;
+        inputDataIds    << gni.qVariableIds;
+        outputDataNames << gni.cVariables;
+        outputDataIds   << gni.cVariableIds;
+        break;
+    }
+    case InterfacePortSpec::MechanicC:
+    {
+        GeneratorNodeInfo gni("NodeMechanic");
+        inputDataNames  << gni.cVariables;
+        inputDataIds    << gni.cVariableIds;
+        outputDataNames << gni.qVariables;
+        outputDataIds   << gni.qVariableIds;
+        break;
+    }
+    case InterfacePortSpec::MechanicRotationalQ:
+    {
+        GeneratorNodeInfo gni("NodeMechanicRotational");
+        inputDataNames  << gni.qVariables;
+        inputDataIds    << gni.qVariableIds;
+        outputDataNames << gni.cVariables;
+        outputDataIds   << gni.cVariableIds;
+        break;
+    }
+    case InterfacePortSpec::MechanicRotationalC:
+    {
+        GeneratorNodeInfo gni("NodeMechanicRotational");
+        inputDataNames  << gni.cVariables;
+        inputDataIds    << gni.cVariableIds;
+        outputDataNames << gni.qVariables;
+        outputDataIds   << gni.qVariableIds;
+        break;
+    }
+    case InterfacePortSpec::HydraulicQ:
+    {
+        GeneratorNodeInfo gni("NodeHydraulic");
+        inputDataNames  << gni.qVariables;
+        inputDataIds    << gni.qVariableIds;
+        outputDataNames << gni.cVariables;
+        outputDataIds   << gni.cVariableIds;
+        break;
+    }
+    case InterfacePortSpec::HydraulicC:
+    {
+        GeneratorNodeInfo gni("NodeHydraulic");
+        inputDataNames  << gni.cVariables;
+        inputDataIds    << gni.cVariableIds;
+        outputDataNames << gni.qVariables;
+        outputDataIds   << gni.qVariableIds;
+        break;
+    }
+    case InterfacePortSpec::PneumaticQ:
+    {
+        GeneratorNodeInfo gni("NodePneumatic");
+        inputDataNames  << gni.qVariables;
+        inputDataIds    << gni.qVariableIds;
+        outputDataNames << gni.cVariables;
+        outputDataIds   << gni.cVariableIds;
+        break;
+    }
+    case InterfacePortSpec::PneumaticC:
+    {
+        GeneratorNodeInfo gni("NodePneumatic");
+        inputDataNames  << gni.cVariables;
+        inputDataIds    << gni.cVariableIds;
+        outputDataNames << gni.qVariables;
+        outputDataIds   << gni.qVariableIds;
+        break;
+    }
+    case InterfacePortSpec::ElectricQ:
+    {
+        GeneratorNodeInfo gni("NodeElectric");
+        inputDataNames  << gni.qVariables;
+        inputDataIds    << gni.qVariableIds;
+        outputDataNames << gni.cVariables;
+        outputDataIds   << gni.cVariableIds;
+        break;
+    }
+    case InterfacePortSpec::ElectricC:
+    {
+        GeneratorNodeInfo gni("NodeElectric");
+        inputDataNames  << gni.cVariables;
+        inputDataIds    << gni.cVariableIds;
+        outputDataNames << gni.qVariables;
+        outputDataIds   << gni.qVariableIds;
+        break;
+    }
+    default:
+        break;
+    }
+
+    foreach(const QString &dataName, inputDataNames)
+    {
+        vars.append(InterfaceVarSpec(dataName, inputDataIds.takeFirst(), InterfaceVarSpec::Input));
+    }
+    foreach(const QString &dataName, outputDataNames)
+    {
+        vars.append(InterfaceVarSpec(dataName, outputDataIds.takeFirst(), InterfaceVarSpec::Output));
+    }
+}
+
+
+InterfaceVarSpec::InterfaceVarSpec(QString dataName, int dataId, InterfaceVarSpec::CausalityEnumT causality)
+{
+    this->dataName = dataName;
+    this->dataId = dataId;
+    this->causality = causality;
+}
+
+void getInterfaces(QList<InterfacePortSpec> &interfaces, hopsan::ComponentSystem *pSystem, QStringList &path)
+{
+    std::vector<hopsan::HString> names = pSystem->getSubComponentNames();
+    for(size_t i=0; i<names.size(); ++i)
+    {
+        hopsan::Component *pComponent = pSystem->getSubComponent(names[i]);
+        hopsan::HString typeName = pComponent->getTypeName();
+
+        if(typeName == "SignalInputInterface")
+        {
+            interfaces.append(InterfacePortSpec(InterfacePortSpec::Input, names[i].c_str(), "out", path));
+        }
+        else if(typeName == "SignalOutputInterface")
+        {
+            interfaces.append(InterfacePortSpec(InterfacePortSpec::Output, names[i].c_str(), "in", path));
+        }
+        else if(typeName == "MechanicInterfaceQ")
+        {
+            interfaces.append(InterfacePortSpec(InterfacePortSpec::MechanicQ, names[i].c_str(), "P1", path));
+        }
+        else if(typeName == "MechanicInterfaceC")
+        {
+            interfaces.append(InterfacePortSpec(InterfacePortSpec::MechanicC, names[i].c_str(), "P1", path));
+        }
+        else if(typeName == "MechanicRotationalInterfaceQ")
+        {
+            interfaces.append(InterfacePortSpec(InterfacePortSpec::MechanicRotationalQ, names[i].c_str(), "P1", path));
+        }
+        else if(typeName == "MechanicRotationalInterfaceC")
+        {
+            interfaces.append(InterfacePortSpec(InterfacePortSpec::MechanicRotationalC, names[i].c_str(), "P1", path));
+        }
+        else if(typeName == "HydraulicInterfaceQ")
+        {
+            interfaces.append(InterfacePortSpec(InterfacePortSpec::HydraulicQ, names[i].c_str(), "P1", path));
+        }
+        else if(typeName == "HydraulicInterfaceC")
+        {
+            interfaces.append(InterfacePortSpec(InterfacePortSpec::HydraulicC, names[i].c_str(), "P1", path));
+        }
+        else if(typeName == "PneumaticInterfaceQ")
+        {
+            interfaces.append(InterfacePortSpec(InterfacePortSpec::PneumaticQ, names[i].c_str(), "P1", path));
+        }
+        else if(typeName == "PneumaticInterfaceC")
+        {
+            interfaces.append(InterfacePortSpec(InterfacePortSpec::PneumaticC, names[i].c_str(), "P1", path));
+        }
+        else if(typeName == "ElectricInterfaceQ")
+        {
+            interfaces.append(InterfacePortSpec(InterfacePortSpec::ElectricQ, names[i].c_str(), "P1", path));
+        }
+        else if(typeName == "ElectricInterfaceC")
+        {
+            interfaces.append(InterfacePortSpec(InterfacePortSpec::ElectricC, names[i].c_str(), "P1", path));
+        }
+        else if(typeName == "Subsystem")
+        {
+            QStringList path2 = path;
+            getInterfaces(interfaces, dynamic_cast<hopsan::ComponentSystem *>(pComponent), path2 << pComponent->getName().c_str());
+        }
+    }
+}
+
+
